prototypes/corp-assistant/pyproject.toml
[project]
name = "seo-marketing"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "aiosqlite>=0.22.1",
    "fastapi[all]>=0.128.0",
    "faststream[redis]>=0.6.5",
    "langchain>=1.2.3",
    "langchain-elasticsearch>=1.0.0",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "markdown-pdf>=1.10",
    "markitdown[all]>=0.1.4",
    "pydub>=0.25.1",
    "python-magic-bin>=0.4.14",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.45",
]

[tool.ruff]
line-length = 99
preview = true
exclude = [
    "migrations",
    "migration",
    "src/migrations",
    "alembic.ini",
    "**/__pycache__",
    "**/.mypy_cache",
    "**/.ruff_cache",
    "**/.venv"
]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    "ASYNC",
    "S",
    "BLE",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "EXE",
    "ISC",
    "ICN",
    "G",
    "INP",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "INT",
    "ARG",
    "TD",
    "FIX",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    "NPY",
    "AIR",
    "PERF",
    "FURB",
    "LOG",
    "RUF",
]
ignore = [
    "S105",
    "S101",
    "B008",
    "TRY201",
    "PLW3201",
    "N805",
    "PTH",
    "COM812",
    "ISC001",
    "TRY003",
    "PT013",
    "TID252",
    "RUF001",
    "RUF002",
    "RUF003",
    "ASYNC109",
]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "typings",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
"typings" = ['typing']

[tool.ruff.lint.pylint]
max-args = 15
max-returns = 10
max-branches = 30

# -- MyPy --
[tool.mypy]
ignore_missing_imports = true
exclude = [
    "src/migration",
    "src/migrations",
    "__samples__",
    ".venv"
]

# -- Vulture --
[tool.vulture]
exclude = [
    ".vscode/",
    "__samples__/",
    "tests",
    ".venv"
]
ignore_names = [
    "post",
    "put",
    "delete",
    "head"
]
